dashboard/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-dotenv==1.0.0
gevent==23.9.1
gevent-websocket==0.10.1

# Web3 and blockchain
web3==6.11.1
eth-account==0.9.0
eth-utils==2.3.0

# System monitoring
psutil==5.9.6
python-engineio==4.8.0
python-socketio==5.10.0

# Database
SQLAlchemy==2.0.23

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
dataclasses-json==0.6.3
